server/requirements.txt
# VE3 Tool Dependencies
# =====================

# Core
pyyaml>=6.0
openpyxl>=3.1.0
requests>=2.31.0

# Image preview
pillow>=10.0.0

# Chrome automation
selenium>=4.15.0
webdriver-manager>=4.0.0
DrissionPage>=4.0.0

# Google Sheets API (for google_login.py)
gspread>=5.12.0
google-auth>=2.23.0

# Utils
websocket-client>=1.6.0
pyautogui>=0.9.54
pyperclip>=1.8.2
pywin32>=306

# 2FA TOTP
pyotp>=2.9.0

# Local Proxy Server
flask>=3.0.0
